Change lambda value to float type for Exponential
Initially, ' Integers to negative integer powers are not allowed.' error would appear when I run the Exponential model. I read the doc file, and it says lam accepts float type only. 

I am not sure whether it is the python version issue or pymc3 version issue. Just want to point out something wrong and the way to fix it on my particular case.