Added the data in zip format
Added the data for Chapter 5 in Zip format, so that it could be
downloaded directly and then extracted in both JuPyter notebook AND
Google Colab. This also makes it easier, regardless of notebook
platofrm, to hit "run all" at the top of the page and run all the
scripts without needing to pause.